+++ packages/help-center/src/components/help-center-step-button.tsx
@@ -1,51 +1,10 @@
 import { HelpCenterSelect } from '@automattic/data-stores';
-import {
-	DIFM_FLOW,
-	DIFM_FLOW_STORE,
-	HUNDRED_YEAR_DOMAIN_FLOW,
-	HUNDRED_YEAR_PLAN_FLOW,
-	WEBSITE_DESIGN_SERVICES,
-} from '@automattic/onboarding';
 import { Button } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { addQueryArgs } from '@wordpress/url';
+import { useFlowCustomOptions, useFlowZendeskUserFields } from '../hooks';
 import { HELP_CENTER_STORE } from '../stores';
 import type { FC } from 'react';
-
-const FLOWS_INITIAL_MESSAGES = {
-	[ DIFM_FLOW ]: 'User is purchasing DIFM plan.',
-	[ DIFM_FLOW_STORE ]: 'User is purchasing DIFM store plan.',
-	[ WEBSITE_DESIGN_SERVICES ]: 'User is purchasing DIFM website design services.',
-	[ HUNDRED_YEAR_PLAN_FLOW ]: 'User is purchasing 100 year plan.',
-	[ HUNDRED_YEAR_DOMAIN_FLOW ]: 'User is purchasing 100 year domain.',
-};
-
-const FLOWS_FLOWNAME = {
-	[ DIFM_FLOW ]: 'messaging_flow_dotcom_difm',
-	[ DIFM_FLOW_STORE ]: 'messaging_flow_dotcom_difm',
-	[ WEBSITE_DESIGN_SERVICES ]: 'messaging_flow_dotcom_difm',
-};
-
-export function getZendeskUserFieldsByFlow( flowName: string ) {
-	const url = window.location.href;
-	let userFieldFlowName = null;
-	let userFieldMessage = null;
-
-	if ( Object.keys( FLOWS_INITIAL_MESSAGES ).includes( flowName ) ) {
-		userFieldMessage = `${
-			FLOWS_INITIAL_MESSAGES[ flowName as keyof typeof FLOWS_INITIAL_MESSAGES ]
-		} URL: ${ url }`;
-	}
-
-	if ( Object.keys( FLOWS_FLOWNAME ).includes( flowName ) ) {
-		userFieldFlowName = FLOWS_FLOWNAME[ flowName as keyof typeof FLOWS_FLOWNAME ];
-	}
-
-	return {
-		userFieldMessage,
-		userFieldFlowName,
-	};
-}
 
 interface HelpCenterButtonProps {
 	helpCenterButtonText?: string;
@@ -63,15 +22,16 @@ const HelpCenterButton: FC< HelpCenterButtonProps > = ( {
 		( select ) => ( select( HELP_CENTER_STORE ) as HelpCenterSelect ).isHelpCenterShown(),
 		[]
 	);
+	const flowCustomOptions = useFlowCustomOptions( flowName || '' );
+	const { userFieldMessage, userFieldFlowName } = useFlowZendeskUserFields( flowName || '' );
 
 	if ( ! helpCenterButtonText ) {
 		return null;
 	}
 
 	function openHelpCenter() {
-		setShowHelpCenter( ! isShowingHelpCenter, hasPremiumSupport );
+		setShowHelpCenter( ! isShowingHelpCenter, hasPremiumSupport, flowCustomOptions );
 		if ( hasPremiumSupport ) {
-			const { userFieldMessage, userFieldFlowName } = getZendeskUserFieldsByFlow( flowName || '' );
 			const urlWithQueryArgs = addQueryArgs( '/odie?provider=zendesk', {
 				userFieldMessage,
 				userFieldFlowName,
