+++ packages/help-center/src/hooks/index.ts
@@ -11,3 +11,5 @@ export { usePostByUrl } from './use-post-by-url';
 export { useContentFilter } from './use-content-filter';
 export { useProductsWithPremiumSupport } from './use-products-with-premium-support';
 export { useProductsCustomOptions } from './use-products-custom-options';
+export { useFlowCustomOptions } from './use-flow-custom-options';
+export { useFlowZendeskUserFields } from './use-flow-zendesk-user-fields';
